fix: further restrict symbol grammar
Even without this, we could get away with single quoted strings in most
cases because of our ordered choice, but I want to conceptually
disallow quotes in symbols so that a quoted string can’t parse as a
symbol in contexts where strings aren’t allowed (like sort fields). I
could do a negative lookahead in the symbol grammar instead to only
disallow leading single quotes, but I don’t see a good reason to have
quotes in the middle of a symbol either – too confusing.