refactor: make operator position in field expressions syntactically determinable
BREAKING CHANGE. Previously, which index of a RawFieldExpression’s list
contained the operator was determined contextually, using information
the parser had about the supported operators. This introduced a ton of
complexity and opened the door to many, nasty bugs in which an argument
(that was an identifier/field reference and happened to share the name
of an operator) could mistakenly get parsed as the operator! That
shouldn’t be able to happen and, similarly, defining new operators
shouldn’t change the meaning of existing expressions — that’s a huge
security risk.

So, this commit instead defines two syntactic forms: one in which the
operator is infixed (the common case) and one in which its the first
item. These are identified by the requirement that there be a  “:”
before the operator name.

This makes `(1,:operator,2)` equivalent to `(:operator,1,2)`. The
latter form, with the leading operator, is strictly more powerful, in
that it can express `(:operator,arg1)` and `(:operator,1,2,3)`, both of
which have no corresponding serializations in the other form. Users can
therefore choose to use the prefix form across the board, or mix and
match the prefix and infix schemes.

This change (requiring the colon) makes the `eq` shorthand, in which
`(arg1,arg2)` is short for `(arg1,:eq,arg2)` safe again, so we keep it.