fix(serialization): symbol serialization bugfix
This commit creates a function for encoding (the string that names) a
symbol that's separate from the function for encoding the contents of a
string literal. This is needed because, for symbols only, we need to
encode certain leading characters that our grammer doesn't allow to
appear literally. Before this change, some symbols were serialized into
values that then couldn’t be parsed. That was annoying, but probably
not a security issue.