URIDecode Query String Before Validating Query
I  tested this on on firefox and postman and it works. Applications such as postman encodeURI before sending it out. This will decode it before evaluating it. I would ask to implement this or create a option to atleast set this if needed. In anycase a simple decode won't hurt the tool. Only make it before at the least.